Add metadata interface for instance data
Some plugins, like the userdata execution plugin, need a standard model
for the instance data, which can be applied by templating engines like
jinja to the userdata.

This way, a user can use this feature to customize her userdata scripts
with values that are specific to that instance runtime env or cloud platform.

The instance data structure is based on the cloud-init specifications:
https://cloudinit.readthedocs.io/en/latest/topics/instancedata.html

The v1 namespace contains a subset of the cloud-init specs for the instance
data.

The ds.meta_data namespace contains all the values the v1 namespace
contains, in order to be compatible with cloud-init, plus a subset
of other instance data.

Change-Id: I5c529498f06fe3c86f7fa3c20fdf3091840c4041